requirements.txt
Flask>=2.3
python-dotenv>=1.0
requests>=2.31
pyserial>=3.0
python-escpos>=3.0
